requirements.txt
azure-cognitiveservices-vision-face
google_images_download
six 
semidbm
pykakasi
Pillow
Flask